Code linking and working on PC, without address issue being fixed and KL25Z without printing

+++ source/mem_test/verify.c
@@ -22,7 +22,7 @@ uint32_t * verify_pattern(uint32_t * loc, size_t length, int8_t seed)
 		}
 		else if (*(temp + i) != pattern_holder[i])
 		{
-			*(buffer_address + j) = (uint64_t) (temp + i);
+			*(buffer_address + j) = (uint32_t) (temp + i);
 			j++;
 		}
 	}
